+++ CodeTester-IntelliJ-IDEA-UITest/src/test/kotlin/de/fxnm/fixtures/CodeTesterToolWindowFixture.kt
@@ -2,23 +2,22 @@ package de.fxnm.fixtures
 
 import com.intellij.remoterobot.RemoteRobot
 import com.intellij.remoterobot.data.RemoteComponent
+import com.intellij.remoterobot.fixtures.CommonContainerFixture
 import com.intellij.remoterobot.fixtures.DefaultXpath
-import com.intellij.remoterobot.fixtures.FixtureName
 import com.intellij.remoterobot.search.locators.byXpath
 import com.intellij.remoterobot.stepsProcessing.step
 import com.intellij.remoterobot.utils.waitFor
 import java.time.Duration
 
-fun RemoteRobot.codeTesterToolWindow(timeout: Duration = Duration.ofSeconds(2),
-                                     function: CodeTesterExplorer.() -> Unit) {
+fun RemoteRobot.codeTesterToolWindow(function: CodeTesterExplorer.() -> Unit) {
     step("CodeTester ToolWindow") {
         find<CodeTesterExplorer>(CodeTesterExplorer::class.java).apply(function)
     }
 }
 
-@DefaultXpath("type","//div[@class='CodeTesterToolWindowPanel']")
+@DefaultXpath("type", "//div[@class='CodeTesterToolWindowPanel']")
 open class CodeTesterExplorer(remoteRobot: RemoteRobot, remoteComponent: RemoteComponent) :
-    DialogFixture(remoteRobot, remoteComponent) {
+    CommonContainerFixture(remoteRobot, remoteComponent) {
 
     fun selectCategory(category: String) {
         val combobox = comboBox(byXpath("//div[@class='ComboBox']"), Duration.ofSeconds(2))