+++ CodeTester-IntelliJ-IDEA-UITest/src/test/kotlin/de/fxnm/fixtures/ResultTreeFixture.kt
@@ -2,6 +2,7 @@ package de.fxnm.fixtures
 
 import com.intellij.remoterobot.RemoteRobot
 import com.intellij.remoterobot.data.RemoteComponent
+import com.intellij.remoterobot.fixtures.CommonContainerFixture
 import com.intellij.remoterobot.fixtures.ContainerFixture
 import com.intellij.remoterobot.fixtures.DefaultXpath
 import com.intellij.remoterobot.search.locators.byXpath
@@ -16,7 +17,7 @@ fun ContainerFixture.checkResultTree(function: ResultTreeFixture.() -> Unit = {}
 
 @DefaultXpath("type", "//div[@class='Tree']")
 class ResultTreeFixture(remoteRobot: RemoteRobot, remoteComponent: RemoteComponent) :
-    ContainerFixture(remoteRobot, remoteComponent) {
+    CommonContainerFixture(remoteRobot, remoteComponent) {
 
     private val tree by lazy {
         find<JTreeFixture>(byXpath("//div[@class='Tree']"), Duration.ofSeconds(10))
@@ -26,7 +27,7 @@ class ResultTreeFixture(remoteRobot: RemoteRobot, remoteComponent: RemoteCompone
         return tree.retrieveData().textDataList.stream().map { s -> s.text }.toList()
     }
 
-    fun doubleClickItem(itemNumber:Int) {
+    fun doubleClickItem(itemNumber: Int) {
         findText(tree.extractData()[itemNumber].text).doubleClick()
     }